+++ library/cartridge_set_control_instance.py
@@ -15,6 +15,69 @@ argument_spec = {
 }
 
 
+GET_TWOPHASE_COMMIT_VERSION_TIMEOUT = 60
+
+
+def get_twophase_commit_versions(control_console, advertise_uris):
+    versions, err = control_console.eval_res_err('''
+        local fiber_clock = require('fiber').clock
+        local pool = require('cartridge.pool')
+
+        local uris, timeout = ...
+
+        local deadline = fiber_clock() + timeout
+
+        local connections = {}
+        for i, uri in ipairs(uris) do
+            local conn, err = pool.connect(uri)
+            if err ~= nil then
+                return nil, tostring(err)
+            end
+
+            table.insert(connections, conn)
+        end
+
+        local futures = {}
+        for _, conn in ipairs(connections) do
+            local future = conn:eval([[
+                local twophase_version = require('cartridge.twophase').VERSION
+                if twophase_version ~= nil then
+                    return twophase_version
+                end
+
+                if rawget(_G, '__cartridge_upload_begin') ~= nil then
+                    return 2
+                else
+                    return 1
+                end
+            ]], nil, {is_async = true})
+
+            table.insert(futures, future)
+        end
+
+        local versions = {}
+
+        for _, future in ipairs(futures) do
+            local wait_timeout = deadline - fiber_clock()
+            if wait_timeout < 0 then
+                wait_timeout = 0
+            end
+
+            local result, err = future:wait_result(wait_timeout)
+            if err ~= nil then
+                return nil, tostring(err)
+            end
+
+            local version = result[1]
+            table.insert(versions, version)
+        end
+
+        return versions
+    ''', advertise_uris, GET_TWOPHASE_COMMIT_VERSION_TIMEOUT)
+
+    return versions, err
+
+
 def get_control_instance(params):
     hostvars = params['hostvars']
     play_hosts = params['play_hosts']
@@ -22,7 +85,7 @@ def get_control_instance(params):
     app_name = params['app_name']
 
     control_console = helpers.get_control_console(console_sock)
-    control_instance_name = None
+    control_instance_candidates = []
 
     members, _ = control_console.eval_res_err('''
         return require('membership').members()
@@ -44,22 +107,32 @@ def get_control_instance(params):
             if instance_name not in hostvars:
                 continue
 
-            control_instance_name = instance_name
-            break
+            control_instance_candidates.append(instance_name)
 
-    if control_instance_name is None:
+    if not control_instance_candidates:
         for instance_name in play_hosts:
             instance_vars = hostvars[instance_name]
             if helpers.is_expelled(instance_vars) or helpers.is_stateboard(instance_vars):
                 continue
 
             if 'replicaset_alias' in instance_vars:
-                control_instance_name = instance_name
-                break
+                control_instance_candidates.append(instance_name)
 
-    if control_instance_name is None:
+    if not control_instance_candidates:
         errmsg = 'Not found any joined instance or instance to create a replicaset'
         return helpers.ModuleRes(failed=True, msg=errmsg)
+
+    advertise_uris = [
+        hostvars[instance_name]['config']['advertise_uri']
+        for instance_name in control_instance_candidates
+    ]
+
+    twophase_commit_versions, err = get_twophase_commit_versions(control_console, advertise_uris)
+    if err is not None:
+        return helpers.ModuleRes(failed=True, msg=err)
+
+    idx = twophase_commit_versions.index(min(twophase_commit_versions))
+    control_instance_name = control_instance_candidates[idx]
 
     # in the ideal imagined world we could just use
     # instance_vars['instance_info'], but if control instance is not